+++ src/routing/osrm_routed_wrapper.cpp
@@ -106,9 +106,25 @@ UserDistance OsrmRoutedWrapper::get_distance_value(
   return utils::round<UserDistance>(matrix_entry.GetDouble());
 }
 
-unsigned
-OsrmRoutedWrapper::get_legs_number(const rapidjson::Value& result) const {
-  return result["routes"][0]["legs"].Size();
+const rapidjson::Value&
+OsrmRoutedWrapper::get_legs(const rapidjson::Value& result) const {
+  assert(result.HasMember("routes") && result["routes"].IsArray() &&
+         !result["routes"].Empty() && result["routes"][0].HasMember("legs") &&
+         result["routes"][0]["legs"].IsArray());
+
+  return result["routes"][0]["legs"];
+}
+
+UserDuration
+OsrmRoutedWrapper::get_leg_duration(const rapidjson::Value& leg) const {
+  assert(leg.HasMember("duration"));
+  return utils::round<UserDuration>(leg["duration"].GetDouble());
+}
+
+UserDistance
+OsrmRoutedWrapper::get_leg_distance(const rapidjson::Value& leg) const {
+  assert(leg.HasMember("distance"));
+  return utils::round<UserDistance>(leg["distance"].GetDouble());
 }
 
 std::string OsrmRoutedWrapper::get_geometry(rapidjson::Value& result) const {