+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 #### Internals
 
+- Bypass matrix request in `plan` mode (#444)
 - Refactor `Matrix` template class (#1089)
 - Refactor to use `std::format` whenever possible (#1081)
 - Reduce complexity for recreation process (#1155)